SLN-112 - Add engineering data 2 field on advanced details page
Currently we have placed the field on dashboard page, we need to remove the field from this page and put it on advanced details page, so that when someone taps on engineering data 2 field on the advanced details page then we open the engineering data 2 edit field page. and when they click save then save the data and go back to advanced details page.